requirements.txt
pyspark
pandas
sqlalchemy
psycopg2-binary
pyarrow
streamlit
plotly